Use native multi-key delete Redis query
The only downside to this change is that passing a list of keys now requires the
calling module to know about the location of the keys. If the command is
executed on a particular Redis instance but one or more of the keys do not exist
there, then those keys simply aren't deleted. Thankfully the return value is a
number representing how many keys were affected, so any number that doesn't
match with the original length of the list means there are keys that either
didn't exist at all or didn't exist on the Redis instance where the command was
executed.